refactor/implement react bulletproof architecture
