+++ src/style/index.css
@@ -1,6 +1,6 @@
 @import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&display=swap');
 @import "tailwindcss";
-@import "../styles/design-tokens.css";
+@import "./design-tokens.css";
 
 :root {
   font-family: var(--font-family-sans);
